backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
pymongo==3.12.0
python-dotenv==0.19.0
bcrypt==3.2.0
